+++ src/api/server/fetch/__tests__/task/createTask.spec.ts
@@ -1,25 +1,23 @@
 import 'whatwg-fetch';
 import { rest } from 'msw';
 import { setupServer } from 'msw/node';
-import { completeTask } from '@/api/client/fetch/task';
+import { createTask } from '@/api/server/fetch/task';
 import {
   InvalidResponseBodyError,
   UnexpectedFeatureError,
-  getDynamicBackendApiUrl,
+  getBackendApiUrl,
 } from '@/features';
 import {
-  mockCompleteTask,
-  mockCompleteTaskUnexpectedResponseBody,
+  mockCreateTask,
+  mockCreateTaskUnexpectedResponseBody,
   mockInternalServerError,
 } from '@/mocks';
 
-const mockHandlers = [
-  rest.patch(getDynamicBackendApiUrl('completeTask', '1'), mockCompleteTask),
-];
+const mockHandlers = [rest.post(getBackendApiUrl('tasks'), mockCreateTask)];
 
 const mockServer = setupServer(...mockHandlers);
 
-describe('src/api/client/fetch/task.ts completeTask TestCases', () => {
+describe('src/api/client/fetch/task.ts createTask TestCases', () => {
   beforeAll(() => {
     mockServer.listen();
   });
@@ -35,53 +33,53 @@ describe('src/api/client/fetch/task.ts completeTask TestCases', () => {
   const mockAppToken =
     'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI5OTk5OTk5OTk5OTk5OTk5OTk5OTk5OTk5OTk5OSIsInByb3ZpZGVyIjoiZ29vZ2xlIiwiZXhwIjoxNjgzNzMxMzIzLCJqdGkiOiIzNTY3ZGIyNy0zM2RlLTQyMTctOGM5Zi01ODhhYjVkMDdhZGQiLCJpYXQiOjE2ODExMzkzOTZ9.wV-4ftbM7EwPvyzoqWTNKaC1eZko3juJ84Q9C6X_dYs';
 
-  it('should be able to complete a task', async () => {
-    const completedTask = await completeTask({
-      taskId: 1,
+  it('should be able to create a task', async () => {
+    const createdTask = await createTask({
+      taskCategoryId: 1,
+      status: 'recording',
+      startAt: '2019-08-24T14:15:22Z',
       appToken: mockAppToken,
     });
 
     const expected = {
       id: 1,
-      status: 'completed',
+      status: 'recording',
       startAt: '2019-08-24T14:15:22Z',
-      endAt: '2019-08-24T18:15:22Z',
-      duration: 14400,
+      endAt: '0000-00-00T00:00:00Z',
+      duration: 0,
       taskCategoryId: 1,
     };
 
-    expect(completedTask).toStrictEqual(expected);
+    expect(createdTask).toStrictEqual(expected);
   });
 
   it('should InvalidResponseBodyError Throw, because unexpected response body', async () => {
     mockServer.use(
-      rest.patch(
-        getDynamicBackendApiUrl('completeTask', '1'),
-        mockCompleteTaskUnexpectedResponseBody
-      )
+      rest.post(getBackendApiUrl('tasks'), mockCreateTaskUnexpectedResponseBody)
     );
 
     const dto = {
-      taskId: 1,
+      taskCategoryId: 1,
+      status: 'recording',
+      startAt: '2019-08-24T14:15:22Z',
       appToken: mockAppToken,
     } as const;
 
-    await expect(completeTask(dto)).rejects.toThrow(InvalidResponseBodyError);
+    await expect(createTask(dto)).rejects.toThrow(InvalidResponseBodyError);
   });
 
-  it('should UnexpectedFeatureError Throw, because http status is not ok', async () => {
+  it('should UnexpectedFeatureError Throw, because http status is not created', async () => {
     mockServer.use(
-      rest.patch(
-        getDynamicBackendApiUrl('completeTask', '1'),
-        mockInternalServerError
-      )
+      rest.post(getBackendApiUrl('tasks'), mockInternalServerError)
     );
 
     const dto = {
-      taskId: 1,
+      taskCategoryId: 1,
+      status: 'recording',
+      startAt: '2019-08-24T14:15:22Z',
       appToken: mockAppToken,
     } as const;
 
-    await expect(completeTask(dto)).rejects.toThrow(UnexpectedFeatureError);
+    await expect(createTask(dto)).rejects.toThrow(UnexpectedFeatureError);
   });
 });
