+++ src/api/server/fetch/__tests__/task/fetchPendingTasks.spec.ts
@@ -1,7 +1,7 @@
 import 'whatwg-fetch';
 import { rest } from 'msw';
 import { setupServer } from 'msw/node';
-import { fetchPendingTasks } from '@/api/client/fetch/task';
+import { fetchPendingTasks } from '@/api/server/fetch/task';
 import {
   InvalidResponseBodyError,
   UnexpectedFeatureError,
